cleanup migration ffi
