+++ discovery_engine/lib/src/ffi/types/migration/data.dart
@@ -14,22 +14,14 @@
 
 import 'dart:ffi';
 
-import 'package:xayn_discovery_engine/src/domain/models/active_data.dart';
-import 'package:xayn_discovery_engine/src/domain/models/document.dart';
-import 'package:xayn_discovery_engine/src/domain/models/view_mode.dart';
 import 'package:xayn_discovery_engine/src/ffi/genesis.ffigen.dart';
 import 'package:xayn_discovery_engine/src/ffi/load_lib.dart' show ffi;
 import 'package:xayn_discovery_engine/src/ffi/types/box.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/date_time.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/document/news_resource.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/document/user_reaction.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/duration.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/embedding.dart';
 import 'package:xayn_discovery_engine/src/ffi/types/list.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/migration_search.dart';
+import 'package:xayn_discovery_engine/src/ffi/types/migration/document.dart';
+import 'package:xayn_discovery_engine/src/ffi/types/migration/search.dart';
 import 'package:xayn_discovery_engine/src/ffi/types/primitives.dart';
 import 'package:xayn_discovery_engine/src/ffi/types/source.dart';
-import 'package:xayn_discovery_engine/src/ffi/types/uuid.dart';
 import 'package:xayn_discovery_engine/src/ffi/types/weighted_source_vec.dart';
 import 'package:xayn_discovery_engine/src/infrastructure/migration.dart';
 
@@ -77,46 +69,6 @@ extension OptionDartMigrationDataFfi on DartMigrationData? {
       final data = self.allocNative();
       ffi.init_option_dart_migration_data_some_at(place, data.move());
     }
-  }
-}
-
-class MigrationDocument {
-  final Document document;
-  final ActiveDocumentData? activeData;
-
-  MigrationDocument(this.document, this.activeData);
-
-  void writeNative(Pointer<RustMigrationDocument> place) {
-    document.documentId.writeNative(ffi.migration_document_place_of_id(place));
-    document.stackId
-        .writeNative(ffi.migration_document_place_of_stack_id(place));
-    // ignore: deprecated_member_use_from_same_package
-    (activeData?.smbertEmbedding)
-        .writeNative(ffi.migration_document_place_of_smbert_embedding(place));
-    document.userReaction
-        .writeNative(ffi.migration_document_place_of_reaction(place));
-    document.resource
-        .writeNative(ffi.migration_document_place_of_resource(place));
-    ffi.init_migration_document_is_active_at(place, document.isActive ? 1 : 0);
-    ffi.init_migration_document_is_searched_at(
-      place,
-      document.isSearched ? 1 : 0,
-    );
-
-    // ignore: deprecated_member_use_from_same_package
-    document.batchIndex
-        .writeNative(ffi.migration_document_place_of_batch_index(place));
-
-    // ignore: deprecated_member_use_from_same_package
-    document.timestamp
-        .writeNative(ffi.migration_document_place_of_timestamp(place));
-
-    (activeData?.viewTime[DocumentViewMode.web])
-        .writeNative(ffi.migration_document_place_of_web_view_time(place));
-    (activeData?.viewTime[DocumentViewMode.reader])
-        .writeNative(ffi.migration_document_place_of_reader_view_time(place));
-    (activeData?.viewTime[DocumentViewMode.story])
-        .writeNative(ffi.migration_document_place_of_story_view_time(place));
   }
 }
